Дописал update операцию, добавил обработку ошибок с выдачей кода 404, добавил логи, документацию, написал первый тест

+++ src/main/java/com/example/onlineShop/service/OnlineShopService.java
@@ -1,33 +1,48 @@
 package com.example.onlineShop.service;
 
 import com.example.onlineShop.dao.PhoneDAO;
-import com.example.onlineShop.dao.impl.PhoneDAOImpl;
 import com.example.onlineShop.entity.PhoneEntity;
 import com.example.onlineShop.model.Phone;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+/**
+ * Сервис для управления телефонами в маркетплейсе.
+ */
 @Service
 public class OnlineShopService {
 
   private final PhoneDAO phoneDAO;
 
+  private final LoggerService loggerService;
+
+  /**
+   * Конструктор сервиса.
+   * @param phoneDAO Объект для доступа к данным телефонов
+   * @param loggerService Сервис для логирования действий
+   */
   @Autowired
-  public OnlineShopService(PhoneDAO phoneDAO) {
+  public OnlineShopService(PhoneDAO phoneDAO, LoggerService loggerService) {
     this.phoneDAO = phoneDAO;
+    this.loggerService = loggerService;
   }
 
+  /**
+   * Получить список всех доступных телефонов.
+   * @return Список объектов типа Phone, представляющих доступные телефоны
+   */
   public List<Phone> allAvailablePhones(){
     List<PhoneEntity> phoneEntities = phoneDAO.get();
+    loggerService.getLoggerReadPhoneEntity();
     List<Phone> phoneList = new ArrayList<>();
     phoneEntities.forEach(phoneEntity -> {
       Phone phone = new Phone();
-      phone.setModelPhone(phoneEntity.getModel());
       List<String> availableStores = new ArrayList<>();
       phoneEntity.getAvailableStoreEntities().forEach(availableStoreEntity -> availableStores.add(availableStoreEntity.getName()));
+      phone.setId(phoneEntity.getId());
+      phone.setModelPhone(phoneEntity.getModel());
       phone.setAvailableStores(availableStores);
       phone.setCost(phoneEntity.getCost());
       phoneList.add(phone);
@@ -35,15 +50,57 @@ public class OnlineShopService {
     return phoneList;
   }
 
+  /**
+   * Добавить новый телефон.
+   * @param phone Объект типа Phone, представляющий новый телефон
+   */
   public void addPhone(Phone phone) {
     phoneDAO.create(phone);
+    loggerService.getLoggerAddPhoneEntity();
+    loggerService.getLoggerAddAvailableStoreEntity();
   }
 
-  public void updateInformationPhone(Phone phone) {
+  /**
+   * Обновить информацию о телефоне.
+   * @param phone Объект типа Phone, представляющий информацию для обновления
+   * @return Объект типа PhoneEntity, обновленный телефон или null, если телефон не найден
+   */
+  public PhoneEntity updateInformationPhone(Phone phone) {
+    PhoneEntity phoneEntityFoundById = findPhoneEntityById(phoneDAO.get(), phone.getId());
+    loggerService.getLoggerReadPhoneEntity();
+    if (phoneEntityFoundById != null) {
+      phoneDAO.update(phone);
+      loggerService.getLoggerUpdatePhoneEntity();
+    }
+    return phoneEntityFoundById;
   }
 
-  public void deletePhone(int idPhone) {
-    phoneDAO.delete(idPhone);
+  /**
+   * Удалить телефон по его идентификатору.
+   * @param idPhone Идентификатор удаляемого телефона
+   * @return Объект типа PhoneEntity, удаленный телефон или null, если телефон не найден
+   */
+  public PhoneEntity deletePhone(int idPhone) {
+    PhoneEntity phoneEntityFoundById = findPhoneEntityById(phoneDAO.get(), idPhone);
+    loggerService.getLoggerReadPhoneEntity();
+    if (phoneEntityFoundById != null) {
+      phoneDAO.delete(idPhone);
+      loggerService.getLoggerDeletePhoneEntity();
+      loggerService.getLoggerDeleteAvailableStoreEntity();
+    }
+    return phoneEntityFoundById;
   }
 
+  /**
+   * Найти телефон по его идентификатору.
+   * @param phoneEntities Список объектов типа PhoneEntity
+   * @param idPhone Идентификатор телефона
+   * @return Объект типа PhoneEntity, найденный телефон или null, если телефон не найден
+   */
+  private PhoneEntity findPhoneEntityById(List<PhoneEntity> phoneEntities, long idPhone) {
+    return phoneEntities.stream()
+      .filter(phoneEntity -> phoneEntity.getId().equals(idPhone))
+      .findFirst()
+      .orElse(null);
+  }
 }
